agents/uncensored_agent.py
"""
Uncensored Agent - Provides uncensored reasoning via local Qwen-14B model
Allows for unrestricted thinking and responses
"""

import os
import logging
from typing import Dict, List, Any, Optional

# Global model instance
qwen_model = None
qwen_tokenizer = None

def initialize_qwen_model(model_path: str) -> bool:
    """
    Initialize the Qwen-14B model
    
    Args:
        model_path: Path to the model
        
    Returns:
        True if successful, False otherwise
    """
    global qwen_model, qwen_tokenizer
    logger = logging.getLogger("Hohenheim.UncensoredAgent")
    
    if qwen_model is not None:
        logger.info("Qwen model already initialized")
        return True
    
    try:
        from transformers import AutoModelForCausalLM, AutoTokenizer
        import torch
        
        logger.info(f"Loading Qwen-14B model from {model_path}")
        
        # Check if model exists
        if not os.path.exists(model_path):
            logger.error(f"Model path does not exist: {model_path}")
            return False
        
        # Load tokenizer and model
        qwen_tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
        
        # Load model with lower precision for efficiency
        qwen_model = AutoModelForCausalLM.from_pretrained(
            model_path,
            device_map="auto",
            trust_remote_code=True,
            torch_dtype=torch.float16  # Use half precision
        )
        
        logger.info("Qwen-14B model loaded successfully")
        return True
        
    except Exception as e:
        logger.error(f"Error initializing Qwen model: {str(e)}")
        return False

def get_uncensored_reasoning(query: str, context: Dict[str, Any] = None) -> Dict[str, Any]:
    """
    Get uncensored reasoning using the local Qwen-14B model
    
    Args:
        query: The query to reason about
        context: Additional context
        
    Returns:
        Reasoning results
    """
    global qwen_model, qwen_tokenizer
    logger = logging.getLogger("Hohenheim.UncensoredAgent")
    
    # Format context
    formatted_context = ""
    if context:
        formatted_context = "\n\nContext:\n"
        for key, value in context.items():
            if key == "relevant_memories" and isinstance(value, list):
                formatted_context += "Relevant Memories:\n"
                for i, memory in enumerate(value, 1):
                    formatted_context += f"  {i}. {memory}\n"
            else:
                formatted_context += f"{key.replace('_', ' ').title()}: {value}\n"
    
    # Check if model is initialized
    if qwen_model is None or qwen_tokenizer is None:
        logger.error("Qwen model not initialized")
        return {
            "error": "Qwen model not initialized",
            "reasoning": "Unable to provide uncensored reasoning because the Qwen-14B model is not loaded."
        }
    
    try:
        import torch
        
        # Prepare prompt
        system_prompt = """You are Hohenheim, an advanced AGI system operating in uncensored mode.
In this mode, you can think freely without restrictions or censorship.
Provide detailed, logical reasoning for the query without holding back.
Be honest, direct, and comprehensive in your analysis."""

        prompt = f"{system_prompt}\n\nQuery: {query}{formatted_context}\n\nReasoning:"
        
        # Generate response
        inputs = qwen_tokenizer(prompt, return_tensors="pt").to(qwen_model.device)
        
        with torch.no_grad():
            outputs = qwen_model.generate(
                inputs["input_ids"],
                max_new_tokens=1024,
                temperature=0.7,
                top_p=0.9,
                do_sample=True
            )
        
        # Decode response
        response_text = qwen_tokenizer.decode(outputs[0][inputs["input_ids"].shape[1]:], skip_special_tokens=True)
        
        return {
            "reasoning": response_text.strip(),
            "source": "qwen-14b-uncensored",
            "importance": 0.9  # High importance for uncensored reasoning
        }
        
    except Exception as e:
        logger.error(f"Error generating uncensored reasoning: {str(e)}")
        return {
            "error": f"Error generating uncensored reasoning: {str(e)}",
            "reasoning": "Unable to provide uncensored reasoning due to an error with the Qwen-14B model."
        }

def download_qwen_model(target_path: str) -> bool:
    """
    Download the Qwen-14B model if not already present
    
    Args:
        target_path: Path to download the model to
        
    Returns:
        True if successful, False otherwise
    """
    logger = logging.getLogger("Hohenheim.UncensoredAgent")
    
    try:
        # Check if model already exists
        if os.path.exists(os.path.join(target_path, "config.json")):
            logger.info(f"Qwen model already exists at {target_path}")
            return True
        
        # Create directory if it doesn't exist
        os.makedirs(target_path, exist_ok=True)
        
        logger.info(f"Downloading Qwen-14B model to {target_path}")
        
        # Use Hugging Face's transformers to download
        from transformers import AutoModelForCausalLM, AutoTokenizer
        
        # Download tokenizer
        tokenizer = AutoTokenizer.from_pretrained("Qwen/Qwen-14B", trust_remote_code=True)
        tokenizer.save_pretrained(target_path)
        
        # Download model
        model = AutoModelForCausalLM.from_pretrained("Qwen/Qwen-14B", trust_remote_code=True)
        model.save_pretrained(target_path)
        
        logger.info("Qwen-14B model downloaded successfully")
        return True
        
    except Exception as e:
        logger.error(f"Error downloading Qwen model: {str(e)}")
        return False